requirements.txt
openai>=1.0.0
pyyaml>=6.0
rich>=13.0
requests>=2.28
beautifulsoup4>=4.11
urllib3>=1.26
